Update patch set 2
Patch Set 2:

The latinized form 'Chronus' isn't used in Germany.
I don't think people will understand that this has something to do with clock if we use 'Chronus'.

I think we should use the 'Chronos' as it is closer to the greek original. But perhaps we should ask our greek translators. ;-)

Patch-set: 2