Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Not sure this works properly.  On my N4 it seems to be OK but on my N10 the weather just would not update, regardless of what I tried.  

May be unrelated though so more testing to be done.  -1'ing for now as well.

Patch-set: 1