Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This is not the best way of doing it.  I am assuming the issue with the font is with certain type of devices - specifically HDPI vs XHDPI?  

A better approach would be to set a font size in dimens.xml for the appropriate display density and in the weather panel reference the font size via the dimens property, that way the font can vary depending on the type of device the widget runs on.

Patch-set: 1