Update patch set 15
Patch Set 15: I would prefer that you didn't submit this

Widget loading, even though optimized in certain ways seems to be slower on initial screen turn on when its a lock screen widget.  The "loading widget" text is showing for a much longer time on the N10 and is sometimes visible on the N4 as well although it happens less.

Seems the problem really presents itself if the widget is not the main lock screen widget but one of the other widgets, it does not refresh as it should.  I'm not sure what the cause is but it cannot be merged as is.  

Danny, I will leave this in your capable hands.

Patch-set: 15