Update patch set 17
Patch Set 17:

Jumped the gun with my last review. I experienced two different instances of the weather widget refusing to update--neither manually, or on it scheduled interval. This new patch has remedied that problem. In efforts not to jump the gun again, I will keep testing. But overall, this looks to be turning out amazingly.

Patch-set: 17