Update patch set 18
Patch Set 18:

David,

why would I joke about this? :)

Nicholas,

if this happens again, please use the latest patch set and enable the DEBUG flag in Constants.java. Reproduce the issue and provide logcat. Thanks.

BTW, the weather data is supposed to be persistent now. Most of the speed increase accounts to this. It's updated either
- when the refresh interval elapses and network is available (if not, it's retried when network gets available)
- when changing the location settings or when touching the weather line in the widget and network is available

Patch-set: 18