Update patch set 1
Patch Set 1: Do not submit

This commit includes two unrelated code changes, one of which is a fix to code not even merged.  I still have an issue with the other commit but apart from that, the fix for the code should be in the other commit, not here.  

Also, keep translations and code separate.

Patch-set: 1