Update patch set 1
Patch Set 1:

My cLock stopped updating weather based on my location over the weekend on d2vzw.  Only way to get weather info now is to manually enter a location, and has to be a major city for it to work.  For NJ, I had to use Trenton (Capital) to finally get some weather data to show up.

Not sure if its yahoo's issue or the roms.  Im going to try a build with this commit and see if anything changes.

Something is definitely wrong right now with cLock/yahoo.

Patch-set: 1