Update patch set 1
Patch Set 1:

Sorry, now I really messed it up. I intented for this change to replace #33251, making #33252 superfluous.

This one; however, includes the the changes from both plus the code style issues fixed (hopefully in the right way; I'm not totally sure regarding the constant assigments with the boolean logic).

My apologies for creating this mess; I'll try to be more careful in the future.

Patch-set: 1