Update patch set 2
Patch Set 2:

Updated the commit message to be more clear and added the attachments to the commit message.

I am not sure about hiding the date for events < 1 week.  I am not always sure what the day's date is and having the date in the event details does not hurt.  Being clear about the date > 1 week out makes a lot of sense.

Patch-set: 2