Update patch set 2
Patch Set 2:

To make it clear: The date is added to all entries except todays, if the _configured_ time interval is longer or equal to one week.

First, I planned to implement like Danny suggested (and I think David thought it was actually).

But on second thought, I thought it would still be confusing on the first look, if the date is missing out on entries, and you are not sure if they occur this week or the next. Therefore, I opted to only leave the date out if by configuration, all entries are only from the current week.

PS: If the question arises: This is already implemented for intervals equaling one week, because with this time interval, entries from the same day of the next week are already shown (so; say its monday, you will see entries of today (monday this week) and monday next week. If there are no other entries, you can not see if from which of both weeks the entries are.

Patch-set: 2