Update patch set 2
Patch Set 2:

You are right, but only for your proposed behavior: when the display of date is based on the actual entry being not in the current week. The solution is different when the display of date is only based on the configured interval being longer than one week. And I think additionally, the 'right' thing would be to make the actual interval begin with the start of day for intervals longer than one day, and in that case, it's a completely different problem. 

So, we first have to decide wether it should be the former or the latter. I vote for depending on the configured interval (if my vote counts), as I implemented in the patch, and as I think David also suggested.

Patch-set: 2