Update patch set 2
Patch Set 2:

I'm not sure I follow the conversation.

IMHO, we should always show the date - it does not hurt and we have enough screen space for it.  I see not real benefit of not showing the date.

If however the date is not shown, it should only not show for events on the current day.

As for the lookahead, it is really only intended to limit the calendar query from a performance perspective.  Whether today + 7 days also returns an event from this hour + 7days, as long as it is a valid event, and the date is shown to avoid confusion, we are good.

Patch-set: 2