Update patch set 2
Patch Set 2:

Thank you for your clarification of the 4 options at hand. That makes discussion much easier :-).

I personally favor option 3 (what my patch implements) midly over option 2. I second Danny's opinion that being more verbose is not as nice (though I'd not say ugly).

But I think that ambiguity should be avoided in any case.
That said, both 3 and 4 premise implicit knowledge by the user to understand what the dates mean:
In 3, he has to know that he set the display interval to less than one week so the entries are always within this range.
In 4, he has to know that entries missing day/month are always within the next week.

Hm. The more I think about this, there more I tend to implement option 2, as this is the only one that is always intelligible to the user.

Patch-set: 2