Update patch set 7
Patch Set 7:

@Danny,  I call it a 'small widget' in the code but really it is a 'summary' widget that contains the weather information when the widget size is too small to contain the full weather info. 

Its more about showing weather information on the minimized widget or not.   Maybe the setting in Weather display should be "Show when minimized" as that really describes what we are doing.   The summary text could say something like "Display summary weather information when the widget is minimized" or something like that.

Patch-set: 7