Update patch set 10
Patch Set 10:

I'm building with it now and will test in a few, looks ready to merge so I will do that soon after.  

The performance issue I mentioned was that I can see the items  being added for a split second - I have a meeting that has long text and for a nano second I can see it at the top end hen it moves down as the others load.  Its not something bad, just something that is new with using a listview and the method you describe above.  

I'm not sure we could or should do much about it and agree that it can be a future patch set if we do.

Thank you to both (Nicolai and Danesh) for adding and perfecting this feature.  You will be making a lot of users very happy.

Patch-set: 10