Update patch set 10
Patch Set 10:

I did not notice that, as I don't use the builtin calendar, but aCalendar (+) instead. I don' see any strange dates in aCalendar, but I can confirm that these appear in the stock calendar for me, too.
As the intent does only open an URI (content://com.android.calendar/events/<eventid>), I assumed the bug to be in the stock calendar app. And in fact, it appears to be a stock calendar bug: when manually opening URIs like the above for existing events (you need to know which id your event has), I can confirm that opening them with the stock calendar yields the same result, while opening them with aCalendar gives me the right thing.

Patch-set: 10