Update patch set 11
Patch Set 11:

Ok, I got it working. I still think this is a bug within the stock calendar we should report there (the official docs don't mention my solution, but you can easily find it on the web as a known issue). In the official docs using just the uri is provided as an example.
Btw I changed the uri creation from manual to builtin.

Patch-set: 11