Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I just tested this, and it looks good to me.

But I found another "glitch" with the calendar display: To test this, I set "calendards to display" to none. This did not work. In getCalendarEvents(), there ist not explicitly set which calendar to request in this case, so it appears to default to one of them, and I still got events displayed.

I'm not sure how to handle this. I think the expected result would be to have no events displayed, on the other hand, it simply doesn't make sense not to select any calendar at all (aside from testing how it would look like :-)). Perhaps the best solution would be to notify the user that at least one calendar has to be selected; OR to provide an option to "select default calendar" (whatever exactly that does), and switch this one on if no calendar is selected.

Patch-set: 1