Update patch set 1
Patch Set 1:

Replying to David: I think this is a valid feature request: on devices with lower resolution/small display, and with long event titles/descriptions, I think all space in the horizontal direction could be required, so removing the columns with the icon would help there.

If you agree, I think with the recent changes it would be simple to update the patch and fix it. We could also consider disabling the icon automatically on low-resolution/small display devices...

Patch-set: 1