Update patch set 4
Patch Set 4:

Thank you for your review.

Yes, I also think that is necessary. I implemented it now in the easiest way I could think of; though, technically only an update of the display of the calendar items is necessary. But I think that would add some complexity to the code.

Patch-set: 4