Update patch set 5
Patch Set 5:

"Next entries" was my name for this feature first. In the middle of imlementation I found "upcoming events" to be a much better name, so changed it to that for all user visible content. But I agree that naming should be consistent, so I changed that all. I'm still not 100% sure if all strings are now properly named (like the calendar_upcoming_events_category).

I'm not sure you noticed the default preference for this. I enabled the feature and enabled the bold display per default, as I think this is unobtrusive enough. Do you agree with that?

Last question for this: Am I correct that translations should go to separate commits?

And for a mostly unrelated idea: It should in principle be possible to apply the approach with the SpannableString() to the styling of the clock numbers, too (bold and typeface changes). I think it is simpler, as it does not involve hiding and unhiding parts of the layout. On the other hand, I see the benefits of having the styling in the styles.xml as a central point to change static styling, so it's just a thought.

Patch-set: 5