Update patch set 1
Patch Set 1:

This seems to cause the small widget on the lock screen to stay in the 'Widget is loading...' state. I haven't debugged this further yet, but I didn't see this on yesterday's build. Expanding and collapsing the (initially collapsed) widget fixes it.
I'll have a look at it tomorrow, if you guys don't have an idea earlier ;)

Patch-set: 1