Update patch set 5
Patch Set 5:

Thank you.  I would hate to see something like this happen again on a major release build in the future.  Is this app available on the market for say release builds to easily update newer versions?  That way if yahoo changes servers again it can be easily fixed?

Patch-set: 5