Update patch set 7
Patch Set 7:

When the weather is showing in minimized view, the text in Dutch on my i9100 gets cut off...for letters like q, p, g, j etc.

See screenshot: https://www.dropbox.com/s/eqsjuif3fa03bak/Screenshot_2013-04-07-18-50-29.png
It should say "Zonnig" but displays "Zonnio"

Should I make a JIRA bug report for this?
(Mostly it's not allowed for the i9100 because we don't have a stable version, but this is a UI issue)

Patch-set: 7