Update patch set 1
Patch Set 1: (3 inline comments)

When just looking at the widget display code, I noticed we do some calculations even if later the corresponding parts are hidden. I think this could be optimized as proposed.

I'm not sure why the weather panel was not hidden when showWeather is set to false in the first place - is there any reason?

Patch-set: 1