Update patch set 3
Patch Set 3: (1 inline comment)

I still think this is an abuse of the API and that a calendar sync adapter integrated into the CardDAV syncer that correctly synchronizes the birthdays is the way to go.

As you mentioned it wouldn't be favourable if social apps like Facebook each sync their own calendar: Would you favour that each app which is interested in birthdays of Facebook contacts directly syncs with Facebook? I certainly wouldn't...

Patch-set: 3