Update patch set 2
Patch Set 2:

Sorry, misread the DPI.....

I agree to not go for option 2)

It's not possible to "lift" the weather text only just by a little? That would be enough and you would not have to make any fonts smaller.
It seems to me the vertical spacing between the temp and the text is a little too much.

Patch-set: 2