Update patch set 1
Patch Set 1:

Yes, it doesn't crash immediately when enabling networking location.

Forgot to mention earlier, it also crashes when clicking "tap to refresh" when no location are set, and it crashes when you enable/disable "use custom location" in weather panel preferences. There was a similar issue with this earlier, but Danny fixed it then.

Patch-set: 1