Update patch set 2
Patch Set 2:

Kyle, let me build with this and see how it works before I comment on it officially.

In principle I don't have an issue with it but I do wonder what effect it would have on the layout of the various widget components in the minimized view on different DPI's - specifically MDPI and HDPI devices.  Screenshots of the layout on those devices would be helpful since I no longer have similar devices to test with myself.

Patch-set: 2