Update patch set 2
Patch Set 2: Code-Review-1

@Jorge, Just noticed the screenshots, thanks for pointing me at them.

@Kyle, looking at it, it does not work with the minimized widget - the widget looks off balance with the weather info on the right and nothing on the left.

For the full widget, it also looks like something is missing with the big gap between the weather and the clock.  Not sure it works aesthetically.

I would not merge it until these items can be addressed in a way that looks balanced/pleasing to the eye - and, given that it would be a lot of work to do, I question whether there is value in doing this at all - users have not been demanding it as far as I know. 

Let me know what you think.  I am -1'ing it for now until we arrive at a conclusion/solution.

Patch-set: 2
Label: Code-Review=-1