Update patch set 2
Patch Set 2:

Thank you both for the feedback, I appreciate it.

I would agree that when weather is enabled, there is clearly a gap between the time and the weather where the date would normally be. And I would also agree that things look off balance when in small widget mode.

The use case I envision (and am using) is when there is nothing other than time being displayed, in which case no gap is visible in small or regular mode, since there is only the time.

My use case seems to be a bit of an edge case, in which case I understand that this change is not necessary or wanted, especially with the gap and off-balance nature of the widget when other settings are enabled.

If you both agree, I will abandon this patch. If I find some better way to address the issues in aesthetics you have pointed out, I will try again in the future.

Let me know if that seems ok to both of you, and thanks again for your feedback. I appreciate your time.

Patch-set: 2
Reviewer: Kyle Kamperschroer <4428@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0