Update patch set 2
Patch Set 2:

@Kyle, firstly, thank you for submitting your code - it is well written and contributions to the codebase is always welcomed and appreciated.  

On the topic of whether you should abandon this or not, what I will say is that in general, users are unforgiving taskmasters - unless the layout issues can be resolved, we will just be opening ourselves to a lot of users complaining about the gap/wasted space between the widget components.  Its unfortunate but, such is the life of a developer :-)

Since this is your first commit on CM, regardless of what happens to this one, I would hope that we see a lot more of you in the future.

Patch-set: 2