Update patch set 1
Patch Set 1: Code-Review-1

Just built with this and IMHO The resulting text is too small for the bulk of devices.   We should figure out what is so special about the Z and do a tailored layout for it if needed.

Patch-set: 1
Reviewer: David van Tonder <1576@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1