Update patch set 1
Patch Set 1:

Prior to this patch, in res/values-hdpi/dimens.xml, the value of widget_big_font_size was 70dip.  On an HDPI device (HTC HD2 using Tytung's unofficial port of CM10.1), the lock screen date was pushed out of the frame by the lock screen time so it wasn't shown.

This patch changed the value to 64dip and the lock screen date is partially drawn inside the frame.  But the bottom third or so is cut off.

Changing the value to 62dip allows the date to be completely shown.

Patch-set: 1
Reviewer: Tony Bommarito <6835@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0