Update patch set 1
Patch Set 1:

I also agree that the use case is sensible. However, I got the feeling the approach might not be optimal; perhaps, some users want it the other way around (clock on homescreen but no appointments and vice versa). It would be neat if each widget could be configured completly independent of the others. That would, however, leave the question how the widgets could be configured after adding them to the screen.

Patch-set: 1