Update patch set 6
Patch Set 6:

Matt, this should do it - it matches your code with some small fixes/additions and some of Danny's stuff as well.  Tests fine on CM10 (4.1.2).  

This will not get merged into CM10.2.  I am leaving it here so you and Danny can review and then I will build the standalone APK and release that on XDA and via the Play store :-).

Danny, once you are OK with this, you can abandon this commit.

Patch-set: 6