Update patch set 11
Patch Set 11:

It's kept time overnight but I'll be more sure by the end of the day since it's too early to tell :D

There was a brief moment when the time was wrong last night after installing. I turned on the screen and the actual time was 11:53 but Chronus said 11:51. However on the next tick Chronus was righted to 11:54. It was only about 30sec or so.

Patch-set: 11
Reviewer: Matt Booth <7325@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0