Update patch set 4
Patch Set 4:

David, I changed the pref strings and fixed the numerous issues you had in CalendarWidgetService. In fact, the code was working on my phone for another ~2 months now since the rebase, but all that issues (and some more i repaired) were re-introduced when I rebase the whole patch for 10.2. I think I need more practice with git and the like.

Regarding the order of events: the CalendarInfo is sorted on adding events, so the anniversaries are also displayed in order.

Patch-set: 4