Update patch set 5
Patch Set 5:

I think I did something wrong; I tried to mimic the way the tick observer is implemented and it appeared to work when first uploading the apk. But now (after a few hours), the ContactObserver does not seem to work any more. So hints from someone with insight are much appreciated.

Patch-set: 5