Update patch set 1
Patch Set 1: Code-Review-2

Steve, this is not an error or diagnostic message, this is an indicator to the user that the weather provider is unreachable and the data cannot be refreshed.  There are many examples of similar informational messages in Android, ios, windows etc and is key to a consistent user experience.  Removing the message will just leave the user in an inconsistent state.

Patch-set: 1