Update patch set 1
Patch Set 1:

I'm not sure why this is a point of contention.

There is absolutely zero case for showing an error message to the user when they initially boot their device just because the network isn't available yet. I'm a developer, and I cringe when I see it every time :) We should show nothing until data is available. The user should not have to tap to refresh anything ever, it should just work.

I have to merge a fix similar to this in some way or another. We can not show error messages to the user as part of their OOBE. Please help me resolve this because the alternative is to disable the widget in the OOBE.

Patch-set: 1