Update patch set 1
Patch Set 1:

It looks perfectly fine in this case Danny.  The clock is visible, and the weather begins showing when it's available. It's not bad UX at all. 

This error can only show up in a few cases:

1. Yahoo is broken and you can't fix it anyway. 

2. The network is down. 

3. Our software is buggy and you should either patch it or switch to something else. 

In the case of #2, a weather widget should not be the primary and most visible way of determining this.

The perception of the user is #3 though.  Logic does not apply here. Any indication that "something isn't right" (especially when the issue will just fix itself anyway) does not instill confidence in the user's mind. 

Read that last bit again.  As engineers, we apply logic to everything. UX doesn't always work that way.  It's highly subjective and based on a huge number of things that we don't think about.

We have received a lot of feedback from outside on this specific issue and we have to fix it. I like this widget a lot and I want it in the OOBE, but I should not have to explain why it's reporting errors all the time.

Patch-set: 1