Update patch set 4
Patch Set 4:

We are reconsidering this submission but it would have to be optional. For several reasons, the lack of source material to make xxhdpi icons is one, these icons will never be the default set but it seems to be a popular set despite not really fitting into the style of Android.

Would you be willing to update this change?

Patch-set: 4
Reviewer: Björn Lundén <16@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0