Update patch set 1
Patch Set 1:

I just looked at a random event in my Google contacts birthday calendar: event is on 9th of November, start and end are distinct and point to 9th of November 0:00 and 10th of November 0:00 respectively. So even if begin == end applies for some all-day events, it's not guaranteed.

Patch-set: 1