Update patch set 2
Patch Set 2:

Not sure why this was merged before one has any chance of addressing the comments, but anyway:

I don't yet understand why one needs an *external* refresh trigger if it's just for the case of 'data queried, but nothing present'. The provider itself can trigger an update in that case, can't it?

Patch-set: 2