Update patch set 3
Patch Set 3:

Yeah, I'm not sure where to put it.  On my device with gps not locked on a location I get at most 5000 m.  However, other devices may be more.

As for weather not being the same across 50 km, I don't necessarily agree.  There's probably a better reason to lower it, and an equally good reason to keep it high.

Patch-set: 3