Update patch set 3
Patch Set 3:

I must add that there is a long time bug in the Yahoo weather conditions... for example, right now, according to the Yahoo official weather provider (from the website), my current weather is 19 degrees celsius, with the top of 20 degrees and bottom of 14 degrees... and according to cLock, in the same place with yahoo again I got 23 degrees current, with top of 24 and bottom of 9! It was never accurate, and sometimes the top temperature is less than the current one! Can somebody take a look about this?

Patch-set: 3