Update patch set 5
Patch Set 5: Code-Review+2

As for the discussion about the default provider, I agree it's annoying that the user cannot change settings after the widget has been created. As a "workaround" of sorts, cLock only has one group of settings so if you try to add another cLock widget and configure it to use a different provider, it'll update the original widget as well.

Patch-set: 5
Label: Code-Review=+2