Update patch set 4
Patch Set 4: Code-Review-1

(8 comments)

I added several comments in-line.  However, regardless, I believe this is ultimately the wrong approach.  It will be much easier to just make the clock panel optional by adding a 'Show clock' switch at the top of the clock settings and then hiding/displaying the clock panel based on the user preference.

This will require much less code and there will be no need to add new providers/services and layouts - all the existing ones can be used with a bit of logic that checks for the show clock pref state and start the tick service etc if the clock is visible.

Patch-set: 4