Update patch set 3
Patch Set 3: Code-Review+2

@Danny Boumann 

Even previous to this patch, I had to unlock to view the weather dialog. I am able to confirm that this patch works to fix the problem on the Nexus 6 and the lock functions as it did prior to the patch. 

If we want to see the dialog without unlocking, someone should implement that properly.

Patch-set: 3