Update patch set 3
Patch Set 3:

> (2 comments)
 > 
 > After one set the background of a widget, is not possible to unset
 > it.

I don't understand this comment...can you elaborate? If I set the background to black, it works, and then I can set to white. I can set it to 100% opaque, and it works, and then I can set it to 0% (completely transparent), which is the original behavior.

Patch-set: 3