LockClock: Ensure configuration screen doesn't crash on orientation
If the user switches orientation when the progress
dialog for custom location is showing, the dialog
is torn down, then the task completes and tries to
dismiss causing a crash.

Since we have multiple dialogs to keep state of
and our activity does not do anything fancy with
orientation change, just override it
and avoid tearing down the ui.

Repro :

1. Click custom location preference for weather
2. Rotate screen after pressing OK
3. Observe crash

issue-id : CYNGNOS-994

Change-Id: Id11ce29ce95324c88f83b135684886acf51e0be2