Update weather data only when the screen is on
Because of Doze and App Standby WeatherUpdateService often fails
to update the weather data. AlarmManager provides methods that
allows alarm to go off even when the device is idling, but they
have some restrictions and Google strongly discourages their use.
Add a screen state listener that enables and disables updates when
the screen is on and off respectively and updates are deferred when
the screen is off.

BUGBASH-198

Change-Id: I1be354003334bf953dbc615e765e663ee7601a2c