Update patch set 7
Patch Set 7:

I couldn't really do anything but monitor the changes in the last few days and I can say they don't work properly. This is because I made a wrong assumption: the listener is registered in the application context and as expected it's removed when locklock gets killed. What I didn't expect is that unlike after a system restart, onEnabled() is not called when the application is restarted, so the listener is not re-added. Solution: go back to the standalone service.

Patch-set: 7
Label: Verified=0