Update patch set 2
Patch Set 2:

> Patch Set 2: Verified+1
> 
> What's up with this? Anything else needed?

It's just the alarm icon, which is white, and can be hard too see on light backgrounds.
I wasn't able to add some shadow to it, unfortunately. But it's not that it was so much better before anyway.
Otherwise this should be good.

Patch-set: 2
Reviewer: Gerrit User 444 <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>